requirements.txt
mysql-connector-python
langdetect
wikipedia
streamlit
sendgrid
requests
pymysql
flask
bs4